requirements.txt
Django==3.2.6
google-cloud-secret-manager==2.7.2
django-guardian>=2.1.0,<2.2.0
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt>=4.7.2,<4.8
django-cors-headers>=3.2.1,<3.3.0
django-rest-passwordreset>=1.1.0,<1.2.0
flake8>=3.7.9,<3.8.0
Pillow>=6.2.1,<6.3.1
psycopg2>=2.8.3,<2.9.0
PyJWT==2.1.0
pytz==2019.3
redis>=3.5.3,<3.6.0
django-redis>=4.12.1,<4.13.0
django-cachalot>=2.2.2,<2.3.0
google-auth>=2.3.0,<2.4.0
google-cloud-storage==1.42.3
google-resumable-media>=2.0.3,<2.1.0
googleapis-common-protos>=1.6.0,<1.7.0
django-storages>=1.8,<1.9
django-stdimage==5.1.1
hashids>=1.3.1,<1.4.0
django-filter==21.1
# protobuf>=3.11.2,<3.12.0
inapppy==2.4.4
django-nested-admin==3.3.3

# Swagger substitute for Django 3
drf-yasg==1.20.0

coverage==5.4
celery==5.1.2
django-environ==0.7.0
gunicorn==20.1.0